Move '/syncer/json_delete_handler.rb' and      '/syncer/json_update_handler.rb'   to '/syncer/uploader/json_delete_handler.rb' and      '/syncer/uploader/json_update_handler.rb'
